Fixed bug with sign-in page allowing undefined users to log in, moved sign-up form into a component.
